swap from mypy to pyright. Do type corrections everywhere except commands modules, because they're going to be swapped to cogs anyway

+++ bot/gameObjects/bounties/bountyBoards/bountyBoardChannel.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING, List, Tuple
-from discord import Embed, HTTPException, Forbidden, NotFound, Client, Message, Colour, channel, File
+from typing import TYPE_CHECKING, Coroutine, List, Tuple
+from discord import Embed, HTTPException, Forbidden, NotFound, Client, Message, Colour, channel, File, TextChannel
 from discord.message import MessageReference
 from PIL import Image, ImageDraw
 from io import BytesIO
@@ -15,6 +15,7 @@ from .... import botState
 import asyncio
 from typing import Any, Awaitable, Callable, Dict, Optional, Protocol, Set, Union, cast
 from .. import solarSystem
+from ....logging import LogCategory
 
 
 stopwatchIcon = 'https://emojipedia-us.s3.dualstack.us-west-1.amazonaws.com/thumbs/120/twitter/259/stopwatch_23f1.png'
@@ -62,7 +63,7 @@ async def deleteMessageWithRetry(message: Message, meta: str, *args, **kwargs):
     :param meta: An extra string to describe the message, only used in exceptions
     :type meta: str
     """
-    return await lib.discordUtil.asyncOperationWithRetry(message.delete, "delete message", "bountyBoards",
+    return await lib.discordUtil.asyncOperationWithRetry(message.delete, "delete message", LogCategory.bountyBoards,
                                                         "BBC", meta, *args, **kwargs)
 
 
@@ -95,8 +96,8 @@ class BountyBoardChannel(Serializable):
     :vartype channel: discord.TextChannel
     """
 
-    def __init__(self, division: "BountyDivision", channelIDToBeLoaded : int, messagesToBeLoaded : Dict[int, dict],
-                    noBountiesMsgToBeLoaded : Union[int, None], escapedBountiesMsgToBeLoaded: Union[int, None]):
+    def __init__(self, division: "BountyDivision", channelIDToBeLoaded: int, messagesToBeLoaded: Dict[int, dict],
+                    noBountiesMsgToBeLoaded: Union[int, None], escapedBountiesMsgToBeLoaded: Union[int, None]):
         """
         :param BountyDivision division: The division that this BBC represents
         :param int channelIDToBeLoaded: The discord channel ID of the channel where this BBC is active,
@@ -118,14 +119,22 @@ class BountyBoardChannel(Serializable):
         # discord message object to list all escaped bounties. Active even when there are no escaped bounties.
         self.escapedBountiesMessage = None
         # discord channel object
-        self.channel = None
+        # This will be None if the BBC has not yet been initialized
+        self._channel: Optional[TextChannel] = None
         # A list of coroutines to await after init is complete
-        self.postInitTasks: Optional[Set[Awaitable]] = None
+        self.postInitTasks: Optional[Set[Coroutine]] = None
         
         self.initialized = False
 
+    
+    @property
+    def channel(self):
+        if not self.initialized:
+            raise lib.exceptions.NotReady("Attempted to access BountyBoardChannel.channel before the BBC was initialized")
+        return cast(TextChannel, self._channel)
 
-    async def makeBountyEmbed(self, bounty : bounty.Bounty) -> Embed:
+
+    async def makeBountyEmbed(self, bounty: bounty.Bounty) -> Embed:
         """Construct a discord.Embed for listing in a bountyBoardChannel
 
         :param Bounty bounty: The bounty to describe in this embed
@@ -143,43 +152,46 @@ class BountyBoardChannel(Serializable):
                 + f"**Bounty Ends:** <t:{int(bounty.endTime)}:R>"
         embed.add_field(name="​", value=infoStr)
 
-        loadoutFieldValue = ""
-        if cfg.bbcShowLoadoutEmojis:
-            loadoutFieldName = "**Loadout:**"
-            weaponsStr = "".join(i.emoji.sendable for i in bounty.activeShip.weapons if i.hasEmoji)
-            modulesStr = "".join(i.emoji.sendable for i in bounty.activeShip.modules if i.hasEmoji)
-            turretsStr = "".join(i.emoji.sendable for i in bounty.activeShip.turrets if i.hasEmoji)
-            
-            statsShown = True
-            if cfg.bbcShowHpDps:
-                totalHp = bounty.activeShip.getArmour() + bounty.activeShip.getShield()
-                duelingStatsStr = f" {totalHp} HP // {bounty.activeShip.getDPS()} DPS"
-                statsShown = False
-
-            if bounty.activeShip.hasEmoji:
-                loadoutFieldValue += f"{bounty.activeShip.emoji.sendable}{'' if statsShown else duelingStatsStr}\n"
+        if bounty.activeShip is not None:
+            loadoutFieldValue = ""
+            if cfg.bbcShowLoadoutEmojis:
+                loadoutFieldName = "**Loadout:**"
+                weaponsStr = "".join(i.emoji.sendable for i in bounty.activeShip.weapons if i.hasEmoji)
+                modulesStr = "".join(i.emoji.sendable for i in bounty.activeShip.modules if i.hasEmoji)
+                turretsStr = "".join(i.emoji.sendable for i in bounty.activeShip.turrets if i.hasEmoji)
+                
                 statsShown = True
+                if cfg.bbcShowHpDps:
+                    totalHp = bounty.activeShip.getArmour() + bounty.activeShip.getShield()
+                    duelingStatsStr = f" {totalHp} HP // {bounty.activeShip.getDPS()} DPS"
+                    statsShown = False
+                else:
+                    duelingStatsStr = ""
+
+                if bounty.activeShip.hasEmoji:
+                    loadoutFieldValue += f"{bounty.activeShip.emoji.sendable}{'' if statsShown else duelingStatsStr}\n"
+                    statsShown = True
+                else:
+                    duelingStatsStr += "\n"
+
+                if weaponsStr:
+                    loadoutFieldValue += f"{weaponsStr}{'' if statsShown else duelingStatsStr}\n"
+                    statsShown = True
+                if modulesStr:
+                    loadoutFieldValue += f"{modulesStr}{'' if statsShown else duelingStatsStr}\n"
+                    statsShown = True
+                if turretsStr:
+                    loadoutFieldValue += f"{turretsStr}{'' if statsShown else duelingStatsStr}\n"
+                    statsShown = True
             else:
-                duelingStatsStr += "\n"
+                loadoutFieldName = "**See the culprit's loadout with:**"
+                if cfg.bbcShowHpDps:
+                    embed.add_field(name="**Dueling stats:**",
+                                    value=f"Total health: {bounty.activeShip.getArmour() + bounty.activeShip.getShield()}\n" \
+                                        + f"Total DPS: {bounty.activeShip.getDPS()}")
 
-            if weaponsStr:
-                loadoutFieldValue += f"{weaponsStr}{'' if statsShown else duelingStatsStr}\n"
-                statsShown = True
-            if modulesStr:
-                loadoutFieldValue += f"{modulesStr}{'' if statsShown else duelingStatsStr}\n"
-                statsShown = True
-            if turretsStr:
-                loadoutFieldValue += f"{turretsStr}{'' if statsShown else duelingStatsStr}\n"
-                statsShown = True
-        else:
-            loadoutFieldName = "**See the culprit's loadout with:**"
-            if cfg.bbcShowHpDps:
-                embed.add_field(name="**Dueling stats:**",
-                                value=f"Total health: {bounty.activeShip.getArmour() + bounty.activeShip.getShield()}\n" \
-                                    + f"Total DPS: {bounty.activeShip.getDPS()}")
-
-        prefix = self.division.owningDB.owningBasedGuild.commandPrefix
-        embed.add_field(name=loadoutFieldName, value=f"{loadoutFieldValue}`{prefix}loadout criminal {bounty.criminal.name}`")
+            prefix = self.division.owningDB.owningBasedGuild.commandPrefix
+            embed.add_field(name=loadoutFieldName, value=f"{loadoutFieldValue}`{prefix}loadout criminal {bounty.criminal.name}`")
 
         embed.set_thumbnail(url=bounty.criminal.icon)
         # embed.add_field(name="**Reward Pool:**", value=lib.stringTyping.commaSplitNum(bounty.reward) + " Credits")
@@ -225,7 +237,7 @@ class BountyBoardChannel(Serializable):
         :return: A jump URL to the identified message
         :rtype: str
         """
-        if self.channel is None:
+        if not self.initialized:
             channelID = self.channelIDToBeLoaded
             guildID = None
         else:
@@ -262,11 +274,11 @@ class BountyBoardChannel(Serializable):
         :return: A message if one is found, None if an error occurred
         :rtype: Optional[Message]
         """
-        if self.channel is None:
+        if self.initialized:
             raise ValueError(f"Attempted loadMessageWithRetry before initializing self.channel")
 
         meta = self.prependJumpUrl(id, logUrls, meta)
-        return await lib.discordUtil.asyncOperationWithRetry(self.channel.fetch_message, "load message", "bountyBoards",
+        return await lib.discordUtil.asyncOperationWithRetry(self.channel.fetch_message, "load message", LogCategory.bountyBoards,
                                                             "BBC", meta, id)
 
 
@@ -282,7 +294,7 @@ class BountyBoardChannel(Serializable):
         if self.channel is None:
             raise ValueError("Attempted to sendMessageWithRetry before initializing self.channel")
 
-        return await lib.discordUtil.asyncOperationWithRetry(self.channel.send, "send message", "bountyBoards",
+        return await lib.discordUtil.asyncOperationWithRetry(self.channel.send, "send message", LogCategory.bountyBoards,
                                                             "BBC", meta, *args, **kwargs)
 
 
@@ -302,7 +314,7 @@ class BountyBoardChannel(Serializable):
             raise ValueError("Attempted to sendMessageWithRetry before initializing self.channel")
 
         meta = self.prependJumpUrl(message.id, logUrls, meta)
-        await lib.discordUtil.asyncOperationWithRetry(message.edit, "edit message", "bountyBoards",
+        await lib.discordUtil.asyncOperationWithRetry(message.edit, "edit message", LogCategory.bountyBoards,
                                                         "BBC", meta, *args, **kwargs)
         return message
 
@@ -320,7 +332,7 @@ class BountyBoardChannel(Serializable):
         return f"g:{self.channel.guild.name}#{self.channel.guild.id} c:{self.channel.name}#{self.channel.id}"
 
 
-    def makeEscapedBountiesMsgKwargs(self, ignoredBounties: Tuple[bounty.Bounty, ...] = ()) -> Dict[str, Union[str, Embed]]:
+    def makeEscapedBountiesMsgKwargs(self, ignoredBounties: Tuple[bounty.Bounty, ...] = ()) -> Dict[str, Union[str, Optional[Embed]]]:
         """Construct an embed listing all escaped bounties in the division.
 
         :return: A kwargs mapping detailing the content of the BBC's escaped bounties message
@@ -360,11 +372,15 @@ class BountyBoardChannel(Serializable):
 
 
     async def _loadEscapedBountiesMessage(self, logUrls: bool):
+        if self.escapedBountiesMsgToBeLoaded is None:
+            raise ValueError("escapedBountiesMsgToBeLoaded not supplied")
         self.escapedBountiesMessage = await self.loadMessageWithRetry(self.escapedBountiesMsgToBeLoaded,
                                                                         "escaped bounties", logUrls)
 
 
     async def _loadNoBountiesMessage(self, logUrls: bool):
+        if self.noBountiesMsgToBeLoaded is None:
+            raise ValueError("noBountiesMsgToBeLoaded not supplied")
         self.noBountiesMessage = await self.loadMessageWithRetry(self.noBountiesMsgToBeLoaded,
                                                                     "no bounties", logUrls)
 
@@ -415,7 +431,7 @@ class BountyBoardChannel(Serializable):
                                                 self.prependJumpUrl(listing.id, logUrls, f"bounty listing for {crim}")))
 
         await tasks.wait()
-        tasks.logExceptions("bountyBoards")
+        tasks.logExceptions(LogCategory.bountyBoards)
         tasks.clear()
         self.escapedBountiesMessage = await self.sendMessageWithRetry(f"escaped bounties {self.guildAndChannelMeta()}",
                                                                         **self.makeEscapedBountiesMsgKwargs())
@@ -433,19 +449,23 @@ class BountyBoardChannel(Serializable):
         elif self.noBountiesMessage is not None:
             self.noBountiesMessage = None
 
-        tasks.logExceptions("bountyBoards")
+        tasks.logExceptions(LogCategory.bountyBoards)
 
 
-    async def init(self, client : Client, logUrls: bool = True):
+    async def init(self, client: Client, logUrls: bool = True):
         """Initialise the BBC's attributes to allow it to function.
         Initialisation is done here rather than in the constructor as initialisation can only be done asynchronously.
 
         :param discord.Client client: A logged in client instance used to fetch the BBC's message and channel instances
         :param logUrls: Whether to include message jump urls in logs (Default True)
         """
-        self.channel = client.get_channel(self.channelIDToBeLoaded) or await client.fetch_channel(self.channelIDToBeLoaded)
-        if self.channel is None:
+        chan = client.get_channel(self.channelIDToBeLoaded) or await client.fetch_channel(self.channelIDToBeLoaded)
+        if chan is None:
             raise lib.exceptions.NoLongerExists(f"Failed to load requested channel: {self.channelIDToBeLoaded}")
+        if not isinstance(chan, TextChannel):
+            raise ValueError(f"Channel is not a TextChannel: {self.channelIDToBeLoaded}")
+        
+        self._channel = chan
 
         tasks = lib.discordUtil.BasicScheduler()
         # True if the channel configuration is invalid and needs to be rebuilt
@@ -471,7 +491,7 @@ class BountyBoardChannel(Serializable):
         # del self.noBountiesMsgToBeLoaded
 
         await tasks.wait()
-        tasks.logExceptions("bountyBoards")
+        tasks.logExceptions(LogCategory.bountyBoards)
 
         if doReload:
             await self.rebuild(logUrls)
@@ -483,7 +503,7 @@ class BountyBoardChannel(Serializable):
                         tasks.add(self._sendBountyMsg(b))
             if tasks:
                 await tasks.wait()
-                tasks.logExceptions("bountyBoards")
+                tasks.logExceptions(LogCategory.bountyBoards)
 
         self.initialized = True
         
@@ -492,19 +512,19 @@ class BountyBoardChannel(Serializable):
             for task in self.postInitTasks:
                 t.add(task)
             await t.wait()
-            t.logExceptions(logCategory="bountyBoards")
+            t.logExceptions(logCategory=LogCategory.bountyBoards)
             del self.postInitTasks
             self.postInitTasks = None
 
     
-    def addPostInitTask(self, coro: Awaitable):
+    def addPostInitTask(self, coro: Coroutine):
         if self.postInitTasks is None:
             self.postInitTasks = {coro}
         else:
             self.postInitTasks.add(coro)
 
 
-    def hasMessageForCriminal(self, criminal : criminal.Criminal) -> bool:
+    def hasMessageForCriminal(self, criminal: criminal.Criminal) -> bool:
         """Decide whether this BBC stores a listing for the given criminal 
 
         :param Criminal criminal: The criminal to check for listing existence
@@ -514,7 +534,7 @@ class BountyBoardChannel(Serializable):
         return criminal in self.bountyMessages
 
 
-    def hasMessageForBounty(self, bounty : bounty.Bounty) -> bool:
+    def hasMessageForBounty(self, bounty: bounty.Bounty) -> bool:
         """Decide whether this BBC stores a listing for the criminal wanted by the given bounty
 
         :param Bounty bounty: The bounty whose criminal to check for listing existence
@@ -524,7 +544,7 @@ class BountyBoardChannel(Serializable):
         return self.hasMessageForCriminal(bounty.criminal)
 
 
-    def getMessageForBounty(self, bounty : bounty.Bounty) -> Message:
+    def getMessageForBounty(self, bounty: bounty.Bounty) -> Message:
         """Return the message acting as a listing for the given bounty's criminal
 
         :param Bounty bounty: The bounty whose criminal to fetch a listing for
@@ -543,7 +563,7 @@ class BountyBoardChannel(Serializable):
         return not bool(self.bountyMessages)
 
 
-    async def addBounty(self, bounty : bounty.Bounty, message : Message, logUrls: bool = True):
+    async def addBounty(self, bounty: bounty.Bounty, message: Message, logUrls: bool = True):
         """Treat the given message as a listing for the given bounty, and store it in the database.
         If the BBC was previously empty, remove the empty bounty board message if one exists.
         If a HTTP error is thrown when attempting to remove the empty board message,
@@ -565,12 +585,12 @@ class BountyBoardChannel(Serializable):
                         + bounty.criminal.name, category='bountyBoards', eventType="LISTING_ADD-EXSTS")
         self.bountyMessages[bounty.criminal] = message
 
-        if removeMsg:
+        if removeMsg and self.noBountiesMessage is not None:
             await deleteMessageWithRetry(self.noBountiesMessage,
                                         self.prependJumpUrl(self.noBountiesMessage.id, logUrls, "no bounties"))
 
 
-    async def removeCriminal(self, criminal : criminal.Criminal):
+    async def removeCriminal(self, criminal: criminal.Criminal):
         """Remove the listing message stored for the given criminal from the database,
         and delete its associated message from discord.
         
@@ -596,7 +616,7 @@ class BountyBoardChannel(Serializable):
             await self._sendNoBountiesMessage()
 
 
-    async def removeBounty(self, bounty : bounty.Bounty):
+    async def removeBounty(self, bounty: bounty.Bounty):
         """Remove the listing message stored for the given bounty from the database. 
         his does not attempt to delete the message from discord.
 
@@ -610,7 +630,7 @@ class BountyBoardChannel(Serializable):
         await self.removeCriminal(bounty.criminal)
 
 
-    async def updateBountyMessage(self, bounty : bounty.Bounty):
+    async def updateBountyMessage(self, bounty: bounty.Bounty):
         """Update the embed for the listing associated with the given bounty.
         This includes newly checked and near-correct systems along the route.
         If a HTTP error is thrown when updating the listing, wait and retry the edit for the number of times defined in cfg
@@ -637,7 +657,7 @@ class BountyBoardChannel(Serializable):
             clearTasks.add(self.removeCriminal(criminal))
         if clearTasks:
             await clearTasks.wait()
-            clearTasks.logExceptions("bountyBoards", "bountyBoardChannel", "clear")
+            clearTasks.logExceptions(LogCategory.bountyBoards, "bountyBoardChannel", "clear")
             await self.updateEscapedBountiesMessage()
 
 
@@ -655,7 +675,7 @@ class BountyBoardChannel(Serializable):
 
 
     @classmethod
-    def deserialize(cls, BBCDict : dict, division: "BountyDivision", **kwargs) -> BountyBoardChannel:
+    def deserialize(cls, BBCDict: dict, division: "BountyDivision", **kwargs) -> BountyBoardChannel:
         """Factory function constructing a new BBC from the information in the provided dictionary
         - the opposite of bountyBoardChannel.serialize
 
@@ -663,8 +683,6 @@ class BountyBoardChannel(Serializable):
         :return: The new bountyBoardChannel object
         :rtype: bountyBoardChannel
         """
-        if BBCDict is None:
-            return None
         return BountyBoardChannel(division, BBCDict["channel"], BBCDict["listings"],
                                     BBCDict.get("noBountiesMsg", -1),
                                     BBCDict.get("escapedBountiesMsg", -1))
